pyproject.toml
[project]
name = "lunch-menu"
version = "0.6.0"
description = "Default template for PDM package"
authors = [
    {name = "khr0106", email = "kanghr990106@gmail.com"},
]
dependencies = ["pandas>=2.2.3", "matplotlib>=3.10.0", "psycopg[binary,pool]>=3.2.4", "streamlit>=1.41.1", "python-dotenv>=1.0.1"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[dependency-groups]
eda = [
    "jupyterlab>=4.3.4",
]
test = [
    "pytest>=8.3.4",
]
